(PC-16827) refacto(identification): rename folder into identificationStart
